Added possibility to configure the NoData response delay.
The file watcher needs to be reinstalled when new configuration is loaded
(there is a new file with new value). For this it is good to abort the old
file watcher, which is possible since Node.js 15.9.0 and greater. Thus
update Node.js and while at it, update also dependencies.

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>

# Conflicts:
#	app/dns-server/nodejs/src/index.js